requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.1
exceptiongroup==1.2.2
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
groq==0.23.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.39
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.5
openai==1.76.2
orjson==3.10.18
packaging==24.2
pandas==2.2.3
peewee==3.18.1
phidata==2.7.10
platformdirs==4.3.7
primp==0.15.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yfinance==0.2.57
zstandard==0.23.0
